feat(ace): slice 8.4 — extract index-signature trust-core module + cross-language golden vectors
Extracts the signed-registry-index sign/verify trio (IndexSignableContent,
canonicalIndexBytes, signIndex, verifyIndexSignature) out of signing.ts into a
dedicated trust-core module tools/ace/index-signature.ts (sibling of canonical.ts
8.1 + package-hash.ts 8.2), and anchors it with a cross-language golden-vector
fixture so a future Rust/F#/C# Ace verifies a signed index byte-identically.

T1 — extract + sweep (move-only, behaviour-preserving):
  - NEW tools/ace/index-signature.ts: the four index symbols; one-way dep on
    signing.ts (keyId/AceSignature/VerifyResult/TrustEntry/RevocationMap) +
    canonical.ts (canonicalBytes) + store.ts (RegistryEntry). No cycle.
  - NEW tools/ace/index-signature.test.ts: index sign/verify unit tests relocated
    from signing.test.ts (round-trip + 4 reason rejections: tampered-content /
    tampered-sig / untrusted-key / unsupported-algo).
  - signing.ts: four index symbols removed (manifest sign/verify + shared
    primitives stay); dropped now-unused RegistryEntry import.
  - Swept every import site (ace.ts / registry-publish.ts / registry-remote.ts /
    registry-revoke.ts + their .test.ts + signing.test.ts) to ./index-signature.ts;
    split imports where a file uses both index + manifest symbols.

T2 — golden-vector fixture (tests/cross-verification/index-signature/):
  - vectors.json: canonical[3] (minimal / with-packages / with-revoked+quarantined,
    each w/ expected_canonical_json) · envelope[1] (fixed PEM →
    expected_key_id=ed25519:0dbe5a2385bc2af4 + expected_sig) · invalid[4]
    (tampered-content + tampered-sig → bad-signature; untrusted-key; unsupported-algo).
  - cross-verify.ts: TS oracle asserts every vector, writes ts-output.json, exits
    non-zero on mismatch. Auto-discovered by tools/ci/cross-verify-all.ts (now 6/6).
  - Envelope sig + key_id + canonical bytes independently re-derived via Python
    `cryptography` at fixture-creation (non-tautological, assert-don't-skip).

Canonical-gate status (operator 2026-06-03): index-signature stays TS-only +
proof-owed → NOT canonical yet. The golden vectors are the data the eventual
4-language byte-lock + proof run on; proof half = formal-coverage / lean-proof lane.

Gates: tsc 0 · bun test tools/ace/ 372 pass / 14 files · cross-verify-all 6/6 ·
all files pure-LF · markdownlint clean · move-only (no semantics change).
