+++ tools/ace/ace.test.ts
@@ -7,7 +7,8 @@ import { parseArgs, main } from "./ace.ts";
 import { listInstalled, contentHash, listTrustedKeys, loadRegistry } from "./store.ts";
 import { readRegistriesConfig } from "./store.ts";
 import { generateKeypair, signManifest } from "./signing.ts";
-import { generateKeypair as gkpA, signIndex as sidxA } from "./signing.ts";
+import { generateKeypair as gkpA } from "./signing.ts";
+import { signIndex as sidxA } from "./index-signature.ts";
 import { packageHash } from "./package-hash.ts";
 import { parseLockfile } from "./lockfile.ts";
 import { parseIndex } from "./registry-remote.ts";
